Make Tekken tokenizer a standalone MistralTekkenTokenizer class
Address review feedback:
- Add MistralTekkenTokenizer(BytePairTokenizer) as a public class instead
  of turning MistralTokenizer into a dual-personality class. MistralTokenizer
  is left untouched (SentencePiece only). convert_tokenizer returns a
  MistralTekkenTokenizer directly when tekken.json is detected.
- Inline the Tekken parsing helpers into convert_mistral.py and remove the
  separate convert_tekken.py, matching how convert_gemma4.py inlines its
  tokenizer helpers.
- Drop mistral-common from requirements-common.txt; it is only needed at
  conversion time and is already lazily imported with an ImportError guard.
