+++ keras_hub/src/models/mistral/mistral_tokenizer_test.py
@@ -1,15 +1,59 @@
-import base64
-import json
 import os
-import tempfile
 
 import pytest
 
+from keras_hub.src.models.mistral.mistral_tokenizer import (
+    MistralTekkenTokenizer,
+)
 from keras_hub.src.models.mistral.mistral_tokenizer import MistralTokenizer
 from keras_hub.src.tests.test_case import TestCase
-from keras_hub.src.utils.transformers.convert_tekken import (
-    convert_tekken_tokenizer,
+
+# A tiktoken-style split pattern, matching the Tekken format.
+_TEKKEN_SPLIT_PATTERN = (
+    r"[^\r\n\p{L}\p{N}]?[\p{Lu}\p{Lt}\p{Lm}\p{Lo}\p{M}]*"
+    r"[\p{Ll}\p{Lm}\p{Lo}\p{M}]+|\p{N}| ?[^\s\p{L}\p{N}]+"
+    r"[\r\n/]*|\s*[\r\n]+|\s+(?!\S)|\s+"
 )
+
+
+def _bytes_to_unicode():
+    bs = (
+        list(range(ord("!"), ord("~") + 1))
+        + list(range(ord("¡"), ord("¬") + 1))
+        + list(range(ord("®"), ord("ÿ") + 1))
+    )
+    cs = bs[:]
+    n = 0
+    for b in range(2**8):
+        if b not in bs:
+            bs.append(b)
+            cs.append(2**8 + n)
+            n += 1
+    return {b: chr(c) for b, c in zip(bs, cs)}
+
+
+def _tekken_init_kwargs():
+    """Build a tiny Tekken (byte-level BPE) vocabulary for offline tests."""
+    byte_encoder = _bytes_to_unicode()
+    # Special tokens occupy the first ids, matching the Tekken layout.
+    special_tokens = ["<unk>", "<s>", "</s>", "<pad>", "[INST]"]
+    vocabulary = {token: i for i, token in enumerate(special_tokens)}
+    offset = len(special_tokens)
+    # The 256 single bytes form the base alphabet.
+    for i in range(256):
+        vocabulary[byte_encoder[i]] = offset + i
+    # A few merges on top.
+    merges = []
+    next_id = offset + 256
+    for a, b in [("t", "h"), ("th", "e"), ("i", "n")]:
+        vocabulary[a + b] = next_id
+        merges.append(f"{a} {b}")
+        next_id += 1
+    return {
+        "vocabulary": vocabulary,
+        "merges": merges,
+        "split_pattern": _TEKKEN_SPLIT_PATTERN,
+    }
 
 
 class MistralTokenizerTest(TestCase):
@@ -28,69 +72,6 @@ class MistralTokenizerTest(TestCase):
             init_kwargs=self.init_kwargs,
             input_data=self.input_data,
             expected_output=[[3, 8, 4, 6], [3, 5, 7, 9]],
-        )
-
-    def test_tekken_tokenizer_basics(self):
-        pytest.importorskip("mistral_common")
-        # Magistral-style Tekken (byte-level BPE) vocabulary.
-        vocab = [
-            {
-                "rank": i,
-                "token_bytes": base64.b64encode(bytes([i])).decode(),
-                "token_str": None,
-            }
-            for i in range(256)
-        ]
-        for rank, piece in [
-            (256, b"th"),
-            (257, b"the"),
-            (258, b"in"),
-            (259, b" t"),
-            (260, b" th"),
-        ]:
-            vocab.append(
-                {
-                    "rank": rank,
-                    "token_bytes": base64.b64encode(piece).decode(),
-                    "token_str": piece.decode("latin-1"),
-                }
-            )
-        tekken_config = {
-            "config": {
-                "pattern": (
-                    r"[^\r\n\p{L}\p{N}]?[\p{Lu}\p{Lt}\p{Lm}\p{Lo}\p{M}]*"
-                    r"[\p{Ll}\p{Lm}\p{Lo}\p{M}]+|\p{N}| ?[^\s\p{L}\p{N}]+"
-                    r"[\r\n/]*|\s*[\r\n]+|\s+(?!\S)|\s+"
-                ),
-                "num_vocab_tokens": 261,
-                "default_vocab_size": 266,
-                "default_num_special_tokens": 5,
-                "version": "v7",
-            },
-            "vocab": vocab,
-            "special_tokens": [
-                {"rank": 0, "token_str": "<unk>", "is_control": True},
-                {"rank": 1, "token_str": "<s>", "is_control": True},
-                {"rank": 2, "token_str": "</s>", "is_control": True},
-                {"rank": 3, "token_str": "<pad>", "is_control": True},
-                {"rank": 4, "token_str": "[INST]", "is_control": True},
-            ],
-        }
-        with tempfile.TemporaryDirectory() as dir_path:
-            path = os.path.join(dir_path, "tekken.json")
-            with open(path, "w") as f:
-                json.dump(tekken_config, f)
-            vocabulary, merges, _, split_pattern = convert_tekken_tokenizer(
-                path
-            )
-        self.run_preprocessing_layer_test(
-            cls=MistralTokenizer,
-            init_kwargs={
-                "vocabulary": vocabulary,
-                "merges": merges,
-                "split_pattern": split_pattern,
-            },
-            input_data=["the tin", "in the"],
         )
 
     def test_errors_missing_special_tokens(self):
@@ -119,3 +100,25 @@ class MistralTokenizerTest(TestCase):
                 preset=preset,
                 input_data=self.input_data,
             )
+
+
+class MistralTekkenTokenizerTest(TestCase):
+    def setUp(self):
+        self.init_kwargs = _tekken_init_kwargs()
+        self.input_data = ["the tin", "in the"]
+
+    def test_tokenizer_basics(self):
+        self.run_preprocessing_layer_test(
+            cls=MistralTekkenTokenizer,
+            init_kwargs=self.init_kwargs,
+            input_data=self.input_data,
+        )
+
+    def test_special_tokens(self):
+        tokenizer = MistralTekkenTokenizer(**self.init_kwargs)
+        self.assertEqual(tokenizer.start_token_id, 1)
+        self.assertEqual(tokenizer.end_token_id, 2)
+        self.assertEqual(tokenizer.pad_token_id, 0)
+        # Round-trip a simple string.
+        output = tokenizer("the tin")
+        self.assertEqual(tokenizer.detokenize(output), "the tin")
