+++ keras_hub/src/utils/transformers/convert_mistral_test.py
@@ -1,3 +1,6 @@
+import base64
+import json
+import os
 import tempfile
 
 import numpy as np
@@ -8,8 +11,66 @@ from keras_hub.src.models.backbone import Backbone
 from keras_hub.src.models.causal_lm import CausalLM
 from keras_hub.src.models.mistral.mistral_backbone import MistralBackbone
 from keras_hub.src.models.mistral.mistral_causal_lm import MistralCausalLM
+from keras_hub.src.models.mistral.mistral_tokenizer import (
+    MistralTekkenTokenizer,
+)
 from keras_hub.src.tests.test_case import TestCase
 from keras_hub.src.utils.transformers import convert_mistral
+
+
+def _write_tekken_file(dir_path):
+    """Write a tiny synthetic `tekken.json` for offline conversion tests."""
+    vocab = [
+        {
+            "rank": i,
+            "token_bytes": base64.b64encode(bytes([i])).decode(),
+            "token_str": None,
+        }
+        for i in range(256)
+    ]
+    for rank, piece in [
+        (256, b"th"),
+        (257, b"the"),
+        (258, b"in"),
+        (259, b" t"),
+        (260, b" th"),
+    ]:
+        vocab.append(
+            {
+                "rank": rank,
+                "token_bytes": base64.b64encode(piece).decode(),
+                "token_str": piece.decode("latin-1"),
+            }
+        )
+    config = {
+        "pattern": (
+            r"[^\r\n\p{L}\p{N}]?[\p{Lu}\p{Lt}\p{Lm}\p{Lo}\p{M}]*"
+            r"[\p{Ll}\p{Lm}\p{Lo}\p{M}]+|\p{N}| ?[^\s\p{L}\p{N}]+"
+            r"[\r\n/]*|\s*[\r\n]+|\s+(?!\S)|\s+"
+        ),
+        "num_vocab_tokens": 261,
+        "default_vocab_size": 266,
+        "default_num_special_tokens": 5,
+        "version": "v7",
+    }
+    special_tokens = [
+        {"rank": 0, "token_str": "<unk>", "is_control": True},
+        {"rank": 1, "token_str": "<s>", "is_control": True},
+        {"rank": 2, "token_str": "</s>", "is_control": True},
+        {"rank": 3, "token_str": "<pad>", "is_control": True},
+        {"rank": 4, "token_str": "[INST]", "is_control": True},
+    ]
+    path = os.path.join(dir_path, "tekken.json")
+    with open(path, "w") as f:
+        json.dump(
+            {
+                "config": config,
+                "vocab": vocab,
+                "special_tokens": special_tokens,
+            },
+            f,
+        )
+    return path
 
 
 class TestTask(TestCase):
@@ -118,3 +179,31 @@ class TestTask(TestCase):
             transformers_config
         )
         self.assertEqual(keras_config["rope_max_wavelength"], 20000.0)
+
+    def test_convert_tekken_tokenizer(self):
+        pytest.importorskip("mistral_common")
+        with tempfile.TemporaryDirectory() as dir_path:
+            path = _write_tekken_file(dir_path)
+            vocabulary, merges, split_pattern = (
+                convert_mistral._convert_tekken_tokenizer(path)
+            )
+        # 5 special tokens + 256 bytes + 5 merges.
+        self.assertEqual(len(vocabulary), 266)
+        self.assertEqual(len(merges), 5)
+        # Special tokens keep their reserved rank as id.
+        self.assertEqual(vocabulary["<unk>"], 0)
+        self.assertEqual(vocabulary["<s>"], 1)
+        self.assertEqual(vocabulary["</s>"], 2)
+        self.assertEqual(vocabulary["<pad>"], 3)
+
+        tokenizer = MistralTekkenTokenizer(
+            vocabulary=vocabulary,
+            merges=merges,
+            split_pattern=split_pattern,
+        )
+        self.assertEqual(tokenizer.start_token_id, 1)
+        self.assertEqual(tokenizer.end_token_id, 2)
+        self.assertEqual(tokenizer.pad_token_id, 0)
+        self.assertEqual(tokenizer.vocabulary_size(), 266)
+        output = tokenizer("the tin")
+        self.assertEqual(tokenizer.detokenize(output), "the tin")
